django-pjt/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
executing==2.1.0
idna==3.10
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
parso==0.8.4
prompt_toolkit==3.0.47
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
